Fix windows build artifact upload
The `windows-latest` runner throws an error when attempting to upload
an artifact to an existing release using the `gh` cli:

    gh release upload <tag>

There seems to be an issue with the MinTTY terminal emulator used on
the windows image. Instead of doing this manually, attempt to use an
established action, softprops/action-gh-release@v2, to achieve the
upload operation.